+++ bind/c-ustring.cpp
@@ -18,22 +18,22 @@ struct ustring {
     ustring(OIIO::string_view str);
 
     CPPMM_IGNORE
-    ustring(const char* str, unsigned long pos, unsigned long n);
+    ustring(const char* str, size_t pos, size_t n);
 
     CPPMM_IGNORE
-    ustring(const char* str, unsigned long n);
+    ustring(const char* str, size_t n);
 
     CPPMM_IGNORE
-    ustring(unsigned long n, char c);
+    ustring(size_t n, char c);
 
     CPPMM_IGNORE
-    ustring(const std::string& str, unsigned long pos, unsigned long n);
+    ustring(const std::string& str, size_t pos, size_t n);
 
     CPPMM_RENAME(copy)
     ustring(const OIIO::ustring& str);
 
     CPPMM_IGNORE
-    ustring(const OIIO::ustring& str, unsigned long pos, unsigned long n);
+    ustring(const OIIO::ustring& str, size_t pos, size_t n);
     ~ustring();
 
     operator OIIO::string_view() const CPPMM_IGNORE;
@@ -43,23 +43,23 @@ struct ustring {
     auto assign(const OIIO::ustring& str) -> const OIIO::ustring&;
 
     CPPMM_IGNORE
-    auto assign(const OIIO::ustring& str, unsigned long pos, unsigned long n)
+    auto assign(const OIIO::ustring& str, size_t pos, size_t n)
         -> const OIIO::ustring&;
 
     CPPMM_IGNORE
     auto assign(const std::string& str) -> const OIIO::ustring&;
 
     CPPMM_IGNORE
-    auto assign(const std::string& str, unsigned long pos, unsigned long n)
+    auto assign(const std::string& str, size_t pos, size_t n)
         -> const OIIO::ustring&;
 
     CPPMM_IGNORE
     auto assign(const char* str) -> const OIIO::ustring&;
 
     CPPMM_IGNORE
-    auto assign(const char* str, unsigned long n) -> const OIIO::ustring&;
+    auto assign(const char* str, size_t n) -> const OIIO::ustring&;
     CPPMM_IGNORE
-    auto assign(unsigned long n, char c) -> const OIIO::ustring&;
+    auto assign(size_t n, char c) -> const OIIO::ustring&;
     CPPMM_IGNORE
     auto assign(OIIO::string_view str) -> const OIIO::ustring&;
     auto operator=(const OIIO::ustring& str) -> const OIIO::ustring&;
@@ -76,9 +76,9 @@ struct ustring {
     CPPMM_IGNORE
     auto string() const -> const std::string&;
     auto clear() -> void;
-    auto length() const -> unsigned long;
-    auto hash() const -> unsigned long;
-    auto size() const -> unsigned long;
+    auto length() const -> size_t;
+    auto hash() const -> size_t;
+    auto size() const -> size_t;
     auto empty() const -> bool;
     CPPMM_IGNORE
     auto begin() const
@@ -95,91 +95,91 @@ struct ustring {
     auto rend() const -> std::reverse_iterator<__gnu_cxx::__normal_iterator<
         const char*, std::__cxx11::basic_string<char>>>;
     CPPMM_IGNORE
-    auto operator[](unsigned long pos) const -> const char&;
+    auto operator[](size_t pos) const -> const char&;
     CPPMM_IGNORE
-    auto copy(char* s, unsigned long n, unsigned long pos) const
-        -> unsigned long;
+    auto copy(char* s, size_t n, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto substr(unsigned long pos, unsigned long n) const -> OIIO::ustring;
+    auto substr(size_t pos, size_t n) const -> OIIO::ustring;
     CPPMM_IGNORE
-    auto find(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto find(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find(const std::string& str, unsigned long pos) const -> unsigned long;
+    auto find(const std::string& str, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find(const char* s, unsigned long pos, unsigned long n) const
-        -> unsigned long;
+    auto find(const char* s, size_t pos, size_t n) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find(const char* s, unsigned long pos) const -> unsigned long;
+    auto find(const char* s, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find(char c, unsigned long pos) const -> unsigned long;
+    auto find(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto rfind(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto rfind(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto rfind(const std::string& str, unsigned long pos) const
-        -> unsigned long;
+    auto rfind(const std::string& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto rfind(const char* s, unsigned long pos, unsigned long n) const
-        -> unsigned long;
+    auto rfind(const char* s, size_t pos, size_t n) const
+        -> size_t;
     CPPMM_IGNORE
-    auto rfind(const char* s, unsigned long pos) const -> unsigned long;
+    auto rfind(const char* s, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto rfind(char c, unsigned long pos) const -> unsigned long;
+    auto rfind(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_first_of(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_first_of(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_of(const std::string& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_first_of(const std::string& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_of(const char* s, unsigned long pos, unsigned long n) const
-        -> unsigned long;
+    auto find_first_of(const char* s, size_t pos, size_t n) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_of(const char* s, unsigned long pos) const -> unsigned long;
+    auto find_first_of(const char* s, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_first_of(char c, unsigned long pos) const -> unsigned long;
+    auto find_first_of(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_last_of(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_last_of(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_of(const std::string& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_last_of(const std::string& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_of(const char* s, unsigned long pos, unsigned long n) const
-        -> unsigned long;
+    auto find_last_of(const char* s, size_t pos, size_t n) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_of(const char* s, unsigned long pos) const -> unsigned long;
+    auto find_last_of(const char* s, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_last_of(char c, unsigned long pos) const -> unsigned long;
+    auto find_last_of(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_first_not_of(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_first_not_of(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_not_of(const std::string& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_first_not_of(const std::string& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_not_of(const char* s, unsigned long pos,
-                           unsigned long n) const -> unsigned long;
+    auto find_first_not_of(const char* s, size_t pos,
+                           size_t n) const -> size_t;
     CPPMM_IGNORE
-    auto find_first_not_of(const char* s, unsigned long pos) const
-        -> unsigned long;
+    auto find_first_not_of(const char* s, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_first_not_of(char c, unsigned long pos) const -> unsigned long;
+    auto find_first_not_of(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
-    auto find_last_not_of(const OIIO::ustring& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_last_not_of(const OIIO::ustring& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_not_of(const std::string& str, unsigned long pos) const
-        -> unsigned long;
+    auto find_last_not_of(const std::string& str, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_not_of(const char* s, unsigned long pos,
-                          unsigned long n) const -> unsigned long;
+    auto find_last_not_of(const char* s, size_t pos,
+                          size_t n) const -> size_t;
     CPPMM_IGNORE
-    auto find_last_not_of(const char* s, unsigned long pos) const
-        -> unsigned long;
+    auto find_last_not_of(const char* s, size_t pos) const
+        -> size_t;
     CPPMM_IGNORE
-    auto find_last_not_of(char c, unsigned long pos) const -> unsigned long;
+    auto find_last_not_of(char c, size_t pos) const -> size_t;
     CPPMM_IGNORE
     auto compare(OIIO::string_view str) const -> int;
     CPPMM_IGNORE
@@ -220,7 +220,7 @@ struct ustring {
     static auto getstats(bool verbose) -> std::string;
 
     CPPMM_IGNORE
-    static auto memory() -> unsigned long;
+    static auto memory() -> size_t;
 
     CPPMM_IGNORE
     static auto make_unique(OIIO::string_view str) -> const char*;
